backend/pyproject.toml
[tool.poetry]
name = "metereolog"
version = "0.0.1"
description = "A web platform to collect weather telemetry data"
authors = [
    "Alisson Claudino",
    "João Pinto Ferreira",
    "Gabriel Mason"
]
license = "GNU GPLv3"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10"
django = "==5.2"
requests = ">=2.32.3,<3.0.0"
django-extensions = "^4.1"
psycopg2 = "^2.9.10"
djangorestframework = "^3.16.0"
django-filter = "^25.1"
markdown = "^3.8"
django-timescaledb = "^0.2.13"
werkzeug = "^3.1.3"

[tool.poetry.group.amd64]
optional = true
[tool.poetry.group.amd64.dependencies]
kombu = ">=5.5.3,<6.0.0"

[tool.poetry.group.arm64]
optional = true
[tool.poetry.group.arm64.dependencies]
kombu = ">=5.5.3,<6.0.0"


[tool.poetry.group.dev]
optional = true
[tool.poetry.group.dev.dependencies]
# developer utils
ipdb = "*"
flake8 = "^7.2.0"
black = "^25.1.0"
isort = "^6.0.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
